将习题注释翻译为中文 finish11

+++ exercises/23_conversions/using_as.rs
@@ -1,11 +1,10 @@
-// Type casting in Rust is done via the usage of the `as` operator.
-// Note that the `as` operator is not only used when type casting. It also helps
-// with renaming imports.
+// 在 Rust 中，类型转换是通过使用 `as` 运算符完成的。
+// 请注意，`as` 运算符不仅在类型转换时使用。它还有助于重命名导入。
 
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    // TODO: Make a conversion before dividing.
-    total / values.len()
+    // TODO: 在除法之前进行转换。
+    total / (values.len() as f64)
 }
 
 fn main() {
